javaDoc bisschen 3 (keine Ahnung von anderen javaDoc)

+++ src/main/java/KAGO_framework/model/abitur/datenstrukturen/Graph.java
@@ -345,7 +345,7 @@ public class Graph{
 	        while (neighbors.hasAccess()) {
 	            Vertex neighbor = neighbors.getContent();
 	            System.out.println("Neighbor: " + neighbor.getID());
-	            int newScore = current.getScore() +1;
+	            double newScore = current.getScore() +1;
 	            
 	            if (newScore < neighbor.getScore()) {
 	                neighbor.setScore(newScore);
